[feat/#324-product-kafka] 파일 move
- 구조 리팩토링
