+++ product-service/src/main/java/com/business/productservice/application/service/v2/ProductServiceImplApiV2.java
@@ -3,11 +3,10 @@ package com.business.productservice.application.service.v2;
 import com.business.productservice.application.dto.v2.request.ReqProductPostDTOApiV2;
 import com.business.productservice.application.dto.v2.request.ReqProductPutDTOApiV2;
 import com.business.productservice.application.dto.v2.response.*;
-import com.business.productservice.application.dto.v3.response.ResStockGetByIdDTOApiV3;
 import com.business.productservice.application.exception.ProductExceptionCode;
 import com.business.productservice.domain.product.entity.ProductEntity;
 import com.business.productservice.domain.product.entity.StockEntity;
-import com.business.productservice.infrastructure.dto.ReqToSlackPostDTOApiV1;
+import com.business.productservice.infrastructure.kafka.dto.ReqToSlackPostDTOApiV1;
 import com.business.productservice.infrastructure.feign.SlackFeignClientApiV1;
 import com.business.productservice.infrastructure.persistence.product.ProductJpaRepository;
 import com.business.productservice.infrastructure.persistence.product.StockJpaRepository;
